Move explore state to a separate file
